Fix ping timeout bug.
Sending messages tolls the server from sending pings, and we only
send a ping if it's been less than 300s since the last time the
server sent us something. So if messages are sent in a way that
both tolls the server's pings and also causes us to wait too long
to send a ping ourselves, the connection is dropped.

This patch tracks the last time we sent a PING, and if it's too
long ago, try that first. Only drop the connection if our PING
subsequently goes unanswered.